fix: cleanup code
