+++ src/components/ConditionalFilter/API/initialState/helpers.ts
@@ -1,24 +1,12 @@
-import { ApolloQueryResult, useApolloClient } from "@apollo/client";
+import { ApolloQueryResult } from "@apollo/client";
 import {
-  _GetChannelOperandsDocument,
   _GetChannelOperandsQuery,
-  _GetChannelOperandsQueryVariables,
-  _SearchAttributeOperandsDocument,
   _SearchAttributeOperandsQuery,
-  _SearchAttributeOperandsQueryVariables,
-  _SearchCategoriesOperandsDocument,
   _SearchCategoriesOperandsQuery,
-  _SearchCategoriesOperandsQueryVariables,
-  _SearchCollectionsOperandsDocument,
   _SearchCollectionsOperandsQuery,
-  _SearchCollectionsOperandsQueryVariables,
-  _SearchProductTypesOperandsDocument,
   _SearchProductTypesOperandsQuery,
-  _SearchProductTypesOperandsQueryVariables,
 } from "@dashboard/graphql";
-import { useState } from "react";
 
-import { FetchingParams } from "../../ValueProvider/TokenArray/fetchingParams";
 import { createOptionsFromAPI } from "../Handler";
 import { InitialState } from "../InitialStateResponse";
 import { InitialAPIResponse } from "./types";
@@ -118,101 +106,3 @@ export const createInitialStateFromData = (
       attribute: {},
     },
   );
-
-export const useDataFromAPI = () => {
-  const client = useApolloClient();
-  const [data, setData] = useState<InitialState>(null);
-  const [loading, setLoading] = useState(true);
-  const queriesToRun: Array<Promise<InitialAPIResponse>> = [];
-
-  const fetchQueries = async ({
-    category,
-    collection,
-    producttype,
-    channel,
-    attribute,
-  }: FetchingParams) => {
-    if (channel.length > 0) {
-      queriesToRun.push(
-        client.query<
-          _GetChannelOperandsQuery,
-          _GetChannelOperandsQueryVariables
-        >({
-          query: _GetChannelOperandsDocument,
-        }),
-      );
-    }
-
-    if (collection.length > 0) {
-      queriesToRun.push(
-        client.query<
-          _SearchCollectionsOperandsQuery,
-          _SearchCollectionsOperandsQueryVariables
-        >({
-          query: _SearchCollectionsOperandsDocument,
-          variables: {
-            collectionsSlugs: collection,
-            first: collection.length,
-          },
-        }),
-      );
-    }
-
-    if (category.length > 0) {
-      queriesToRun.push(
-        client.query<
-          _SearchCategoriesOperandsQuery,
-          _SearchCategoriesOperandsQueryVariables
-        >({
-          query: _SearchCategoriesOperandsDocument,
-          variables: {
-            categoriesSlugs: category,
-            first: category.length,
-          },
-        }),
-      );
-    }
-
-    if (producttype.length > 0) {
-      queriesToRun.push(
-        client.query<
-          _SearchProductTypesOperandsQuery,
-          _SearchProductTypesOperandsQueryVariables
-        >({
-          query: _SearchProductTypesOperandsDocument,
-          variables: {
-            productTypesSlugs: producttype,
-            first: producttype.length,
-          },
-        }),
-      );
-    }
-
-    if (Object.keys(attribute).length > 0) {
-      queriesToRun.push(
-        client.query<
-          _SearchAttributeOperandsQuery,
-          _SearchAttributeOperandsQueryVariables
-        >({
-          query: _SearchAttributeOperandsDocument,
-          variables: {
-            attributesSlugs: Object.keys(attribute),
-            choicesIds: Object.values<string[]>(attribute).flat(),
-            first: Object.keys(attribute).length,
-          },
-        }),
-      );
-    }
-
-    const data = await Promise.all(queriesToRun);
-    const initialState = createInitialStateFromData(data, channel);
-    setData(initialState);
-    setLoading(false);
-  };
-
-  return {
-    data,
-    loading,
-    fetchQueries,
-  };
-};
