+++ src/components/ConditionalFilter/API/initialState/index.ts
@@ -1,1 +1,1 @@
-export * from "./useInitalAPIState";
+export * from "./useInitialAPIState";
