+++ src/components/ConditionalFilter/ValueProvider/useUrlValueProvider.ts
@@ -1,8 +1,8 @@
-import useNavigator from "@dashboard/hooks/useNavigator";
 import { stringify } from "qs";
 import { useEffect } from "react";
 import useRouter from "use-react-router";
 
+import { InitialAPIState } from "../API";
 import { FilterContainer } from "../FilterElement";
 import { FilterValueProvider } from "../FilterValueProvider";
 import { useTokenArray } from "./TokenArray";
@@ -23,20 +23,22 @@ const prepareStructure = (filterValue: FilterContainer): Structure =>
     return f.asUrlEntry();
   });
 
-export const useUrlValueProvider = (initalState: any): FilterValueProvider => {
+export const useUrlValueProvider = (
+  initialState: InitialAPIState,
+): FilterValueProvider => {
   const router = useRouter();
   const params = new URLSearchParams(router.location.search);
-  const navigate = useNavigator();
+  const { data, loading, fetchQueries } = initialState;
 
   params.delete("asc");
   params.delete("sort");
 
   const tokenizedUrl = useTokenArray(params.toString());
   const fetchingParams = tokenizedUrl.getFetchingParams();
-  const { data, loading } = initalState;
   useEffect(() => {
-    initalState.fetchQueries(fetchingParams);
+    fetchQueries(fetchingParams);
   }, []);
+
   const value = loading ? [] : tokenizedUrl.asFilterValuesFromResponse(data);
 
   const persist = (filterValue: FilterContainer) => {
