+++ src/components/ConditionalFilter/context/provider.tsx
@@ -1,9 +1,7 @@
 import React, { FC } from "react";
 
-import {
-  useInitialState,
-  useProductFilterAPIProvider,
-} from "../API/ProductFilterAPIProvider";
+import { useProductInitialAPIState } from "../API/initialState/useInitialAPIState";
+import { useProductFilterAPIProvider } from "../API/ProductFilterAPIProvider";
 import { useContainerState } from "../useContainerState";
 import { useFilterLeftOperandsProvider } from "../useFilterLeftOperands";
 import { useUrlValueProvider } from "../ValueProvider/useUrlValueProvider";
@@ -11,12 +9,10 @@ import { ConditionalFilterContext } from "./context";
 
 export const ConditionalProductFilterProvider: FC = ({ children }) => {
   const apiProvider = useProductFilterAPIProvider();
-  const initalState = useInitialState();
-  const valueProvider = useUrlValueProvider(initalState);
+  const initialState = useProductInitialAPIState();
+  const valueProvider = useUrlValueProvider(initialState);
   const leftOperandsProvider = useFilterLeftOperandsProvider();
-  const { value, updateAt, removeAt, createEmpty } = useContainerState(
-    valueProvider.value,
-  );
+  const containerState = useContainerState(valueProvider.value);
 
   return (
     <ConditionalFilterContext.Provider
@@ -24,12 +20,7 @@ export const ConditionalProductFilterProvider: FC = ({ children }) => {
         apiProvider,
         valueProvider,
         leftOperandsProvider,
-        state: {
-          value,
-          updateAt,
-          removeAt,
-          createEmpty,
-        },
+        containerState,
       }}
     >
       {children}
