+++ src/products/views/ProductList/ProductList.tsx
@@ -375,8 +375,6 @@ export const ProductList: React.FC<ProductListProps> = ({ params }) => {
     queryString: params,
   });
 
-  console.log("product list");
-
   return (
     <PaginatorContext.Provider value={paginationValues}>
       <ProductListPage
